refactor: reduce memory usage of solver (#7)

+++ src/main.rs
@@ -27,8 +27,8 @@ enum Commands {
     /// Solve the current game
     Solve {
         /// Maximum number of states to explore
-        #[arg(long, default_value_t = 50_000_000, value_name = "NUM")]
-        max_states: usize,
+        #[arg(long, default_value_t = 100_000_000, value_name = "NUM")]
+        max_states: u32,
         /// Return early when any solution is found (not necessarily minimal)
         #[arg(long)]
         fast: bool,
@@ -39,8 +39,8 @@ enum Commands {
     #[cfg(windows)]
     Autoplay {
         /// Maximum number of states to explore
-        #[arg(long, default_value_t = 50_000_000, value_name = "NUM")]
-        max_states: usize,
+        #[arg(long, default_value_t = 100_000_000, value_name = "NUM")]
+        max_states: u32,
         /// Return early when any solution is found (not necessarily minimal)
         #[arg(long)]
         fast: bool,
@@ -96,7 +96,7 @@ fn main() -> Result<()> {
     Ok(())
 }
 
-fn do_solve(board: Board, max_states: usize, minimal: bool) -> Result<Vec<Action>> {
+fn do_solve(board: Board, max_states: u32, minimal: bool) -> Result<Vec<Action>> {
     let state = board.pretty_print();
     let SolveResult {
         actions,
